# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI CNC Manufacturing Intelligence Platform",
    "description": "\nAI CNC Manufacturing Intelligence Platform.\n\nAn engineering copilot, not a machine controller. Photo-derived geometry is\nprovisional, deterministic services perform every engineering calculation, and a\nqualified engineer remains accountable for dimensions, workholding, tooling,\npostprocessing and final release.\n\n**Safety model**\n\n* Every manufacturing attribute stores value, unit, source, confidence and verification status.\n* Verified CAD and PMI outrank drawings, measurements, scans, calibrated photographs, ordinary photographs and AI inference.\n* Geometry, kinematics, collision, cutting parameters and postprocessing are deterministic.\n* Only a user with release authority may approve a validated machine and post pair.\n* Missing critical data produces a stop or a verification task, never a silent invention.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/geometry-jobs": {
      "post": {
        "tags": [
          "geometry"
        ],
        "summary": "Request a reconstruction",
        "description": "Queue reconstruction against immutable artifact ids.\n\nEngineering calculations are asynchronous; the job exposes stage, percent\nand diagnostics while it runs (PRD 9.1).",
        "operationId": "create_geometry_job_v1_geometry_jobs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeometryJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/geometry": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "List geometry revisions",
        "operationId": "list_geometry_v1_projects__project_id__geometry_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeometryOut"
                  },
                  "title": "Response List Geometry V1 Projects  Project Id  Geometry Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "Read one geometry revision",
        "operationId": "get_geometry_v1_geometry__geometry_id__get",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeometryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}/confidence-map": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "Confidence map for engineering review",
        "description": "Every value with its source, method, confidence, residual and verifier.",
        "operationId": "confidence_map_v1_geometry__geometry_id__confidence_map_get",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confidence Map V1 Geometry  Geometry Id  Confidence Map Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}/surface": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "Finished-part surface for the 3D viewer",
        "description": "The finished surface as a height field, sampled on the engine's own grid.\n\nThe viewer draws exactly what the verification engine compares against,\nrather than a separate cosmetic model that could drift from it.",
        "operationId": "geometry_surface_v1_geometry__geometry_id__surface_get",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "pitch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Pitch"
            }
          },
          {
            "name": "downsample",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "title": "Downsample"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Geometry Surface V1 Geometry  Geometry Id  Surface Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}/features": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "Recognised features",
        "operationId": "list_features_v1_geometry__geometry_id__features_get",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Features V1 Geometry  Geometry Id  Features Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "geometry"
        ],
        "summary": "Edit features into a new revision",
        "description": "Apply feature edits.\n\nThe edit never overwrites: it produces revision N+1 and marks everything\nbuilt from N as stale (PRD 8.2).",
        "operationId": "edit_features_v1_geometry__geometry_id__features_post",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeometryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}:approve": {
      "post": {
        "tags": [
          "geometry"
        ],
        "summary": "Approve geometry for planning",
        "description": "Runs the confidence policy. A critical unknown blocks the approval.",
        "operationId": "approve_geometry_v1_geometry__geometry_id__approve_post",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeometryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/geometry/{geometry_id}/gate": {
      "get": {
        "tags": [
          "geometry"
        ],
        "summary": "Evaluate the geometry approval gate without signing",
        "operationId": "geometry_gate_v1_geometry__geometry_id__gate_get",
        "parameters": [
          {
            "name": "geometry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Geometry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Geometry Gate V1 Geometry  Geometry Id  Gate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "type": "object",
        "title": "ApprovalRequest"
      },
...
      "GeometryJobRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "artifact_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Artifact Ids"
          },
          "known_dimensions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Known Dimensions"
          },
          "parent_geometry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Geometry Id"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "GeometryJobRequest"
      },
      "GeometryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "units": {
            "type": "string",
            "title": "Units"
          },
          "scale_established": {
            "type": "boolean",
            "title": "Scale Established"
          },
          "scale_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Source"
          },
          "scale_uncertainty_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Uncertainty Mm"
          },
          "provisional": {
            "type": "boolean",
            "title": "Provisional"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "quality_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality Report"
          },
          "part_model": {
            "additionalProperties": true,
            "type": "object",
            "title": "Part Model"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "revision",
          "parent_id",
          "units",
          "scale_established",
          "scale_source",
          "scale_uncertainty_mm",
          "provisional",
          "status",
          "content_hash",
          "quality_report",
          "part_model",
          "created_at",
          "version"
        ],
        "title": "GeometryOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "percent": {
            "type": "number",
            "title": "Percent"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "project_id",
          "status",
          "stage",
          "percent",
          "payload",
          "result",
          "error",
          "attempts",
          "trace_id",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}